Added Multi-domain-sentiment baseline results with svm linear kernel using scikit learn and also added multi-domain-sentiment svmlight format generator

+++ research/nltk/sentiment_example.py
@@ -11,8 +11,8 @@ posids = movie_reviews.fileids('pos')
 negfeats = [(word_feats(movie_reviews.words(fileids=[f])), 'neg') for f in negids]
 posfeats = [(word_feats(movie_reviews.words(fileids=[f])), 'pos') for f in posids]
  
-negcutoff = len(negfeats)*3/4
-poscutoff = len(posfeats)*3/4
+negcutoff = int(len(negfeats)*3/4)
+poscutoff = int(len(posfeats)*3/4)
  
 trainfeats = negfeats[:negcutoff] + posfeats[:poscutoff]
 testfeats = negfeats[negcutoff:] + posfeats[poscutoff:]
